EESB-444 Move `TableHeadersConfig` type out of getDefaultTableHeadersConfig.ts
