+++ src/explore-education-statistics-common/src/modules/table-tool/types/permalink.ts
@@ -1,6 +1,6 @@
+import { TableHeadersConfig } from '@common/modules/table-tool/types/tableHeaders';
 import { TableDataQuery } from '@common/services/tableBuilderService';
 import { FullTable } from '@common/modules/table-tool/types/fullTable';
-import { TableHeadersConfig } from '@common/modules/table-tool/utils/tableHeaders';
 
 interface PermalinkQuery extends TableDataQuery {
   configuration: {
