+++ src/explore-education-statistics-common/src/modules/table-tool/types/tableHeaders.ts
@@ -1,0 +1,8 @@
+import { Filter } from '@common/modules/table-tool/types/filters';
+
+export interface TableHeadersConfig {
+  columns: Filter[];
+  columnGroups: Filter[][];
+  rows: Filter[];
+  rowGroups: Filter[][];
+}
