+++ src/explore-education-statistics-frontend/src/modules/permalink/PermalinkPage.tsx
@@ -7,7 +7,7 @@ import permalinkService, {
 } from '@common/services/permalinkService';
 import mapPermalink from '@common/modules/table-tool/utils/mapPermalink';
 import mapTableHeadersConfig from '@common/modules/table-tool/utils/mapTableHeadersConfig';
-import getDefaultTableHeaderConfig from '@common/modules/table-tool/utils/tableHeaders';
+import getDefaultTableHeaderConfig from '@common/modules/table-tool/utils/getDefaultTableHeadersConfig';
 import ButtonLink from '@frontend/components/ButtonLink';
 import Page from '@frontend/components/Page';
 import PrintThisPage from '@frontend/components/PrintThisPage';
