+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/TableTool.tsx
@@ -11,9 +11,9 @@ import TableToolWizard, {
 import TimePeriodDataTable from '@common/modules/table-tool/components/TimePeriodDataTable';
 import WizardStep from '@common/modules/table-tool/components/WizardStep';
 import WizardStepHeading from '@common/modules/table-tool/components/WizardStepHeading';
+import { TableHeadersConfig } from '@common/modules/table-tool/types/tableHeaders';
 import permalinkService from '@common/services/permalinkService';
 import { FullTable } from '@common/modules/table-tool/types/fullTable';
-import { TableHeadersConfig } from '@common/modules/table-tool/utils/tableHeaders';
 import { OmitStrict } from '@common/types';
 import Link from '@frontend/components/Link';
 import React, { useEffect, useRef, useState } from 'react';
